Backend/Controller/requirements.txt
absl-py==0.15.0
amqp==5.0.6
asgiref==3.4.1
astunparse==1.6.3
billiard==3.6.4.0
cachetools==4.2.4
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-prometheus==2.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
docx==0.2.4
flatbuffers==2.0
flower==1.0.0
gast==0.4.0
gevent==21.8.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.41.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.5.0
humanize==3.12.0
idna==3.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kombu==5.2.0
libclang==12.0.0
lxml==4.6.4
Markdown==3.3.4
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.4
pdf2image==1.16.0
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyPDF2==1.26.0
pytesseract==0.3.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.1
pytz==2021.1
rabbitmq==0.2.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
threadpoolctl==3.0.0
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
xgboost==1.5.0
zope.event==4.5.0
zope.interface==5.4.0
